Skip updating stack config if it is detached from a display
Ensuring visibility or configuration of activities in a stack can be
triggered from outside of system service by a client callback, e.g.
ActivityManagerService#activityPaused(). We should skip doing that
whenever we see that a stack is no longer attached to a display.

This CL restores the behavior that was in place where the error might
happen before switching to using
ActivityStackSupervisor#ensureVisibilityAndConfig() and will only
ensure system-wide activity visibility without updating config.

Bug: 110074206
Test: Build and run
Change-Id: Ib347b56b66ce27e4ed8c263ffb425d336012a4ff