Adjusted tracing of reportFullyDrawn() API call.
This patch adjusts how the reportFullyDrawn() API call records this
information in the tracing system.  Instead of an asynch slice that may
or may not be ended by reportFullyDrawn() we now emit a slice around the
function used to handle this callback.  This slice name now also
includes the name of the package that made the callback.

Test: Built and loaded the image; launched Photos
Change-Id: I8d55be73303133a8a1e4648b1fc9eb7938a453ab