[GWP-ASan] Ensure recoverable crashes show up in AppExitInfo.
ActivityManager's getHistoricalProcessExitReasons API is important for
allowing an app to access its own native crashes. For that reason, it
should support GWP-ASan recoverable crashes.

Currently, because recoverable crashes don't actually cause an app
crash, the tombstones are not available. The other problem with just
naively shoving them into the
AppExitInfoTracker::AppExitInfoContainer::mInfos is that it's an API
that only supports a single exitinfo for each pid. So, we'd end up
causing a recoverable crash, then the app would be closed regularly, and
that would overwrite the recoverable crash. So, this patch adds a
separate stash for storing the recoverable crashes.

Bug: 270720163
Test: atest CtsGwpAsanTestCases
Change-Id: I9418fa9e40bdca03516da2b387c13d59fbdf182b